hitl_ui/requirements.txt
Flask==2.3.3
firebase-admin==6.0.1
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.33.0
google-cloud-secret-manager==2.16.4
requests==2.31.0
